Avoid timeouts while publishing a package on Pypi
Publishing packages on PyPI using GitHub actions
can take longer than expected. Poetry has a timeout
of 15 seconds to check for new activity on the network
but on GitHub actions sometimes that's not enough.
This error was reported on
https://github.com/python-poetry/poetry/issues/6009
and it's fixed using the env variable POETRY_REQUESTS_TIMEOUT.
We set this variable to 120 seconds that should be
more that enough for PyPI to finish the work.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>